[QFEED-158] Refactor: 질문 생성 로직 리팩토링
