+++ module-domain/src/main/java/com/wsws/moduledomain/feed/question/ai/QuestionGenerateClient.java
@@ -8,7 +8,7 @@ public interface QuestionGenerateClient {
     /**
      * 생성해야하는 카테고리 목록들을 받아 프롬프트를 작성해 질문 생성
      */
-    Map<String, String> createQuestions(List<String> categories, Map<String, Set<String>> questionBlackListMap);
+    Map<String, String> generateQuestions(List<String> categories, Map<String, Set<String>> questionBlackListMap);
 
     /**
      * 토큰 계산 결과 받아오기
